requirements.txt
Pillow==9.5.0
numpy==1.24.3
pyserial==3.5